+++ packages/test-base/src/androidAndroidTest/kotlin/io/realm/kotlin/test/shared/notifications/RealmListNotificationsTests.kt
@@ -301,9 +301,6 @@ class RealmListNotificationsTests : RealmEntityNotificationTests {
     @Test
     override fun cancelAsFlow() {
         runBlocking {
-            // Freeze values since native complains if we reference a package-level defined variable
-            // inside a write block
-            val values = OBJECT_VALUES.freeze()
             val container = realm.write {
                 copyToRealm(RealmListContainer())
             }
@@ -331,7 +328,7 @@ class RealmListNotificationsTests : RealmEntityNotificationTests {
             // Trigger an update
             realm.write {
                 val queriedContainer = findLatest(container)
-                queriedContainer!!.objectListField.addAll(values)
+                queriedContainer!!.objectListField.addAll(OBJECT_VALUES)
             }
             assertEquals(OBJECT_VALUES.size, channel1.receive().list.size)
             assertEquals(OBJECT_VALUES.size, channel2.receive().list.size)